+++ swing/framework-model/src/main/java/is/codion/swing/framework/model/SwingEntityEditor.java
@@ -18,17 +18,12 @@
  */
 package is.codion.swing.framework.model;
 
-import is.codion.common.model.component.combobox.FilterComboBoxModel;
-import is.codion.common.utilities.proxy.ProxyBuilder;
 import is.codion.framework.db.EntityConnectionProvider;
 import is.codion.framework.domain.entity.Entity;
 import is.codion.framework.domain.entity.EntityDefinition;
 import is.codion.framework.domain.entity.EntityType;
-import is.codion.framework.domain.entity.attribute.Attribute;
 import is.codion.framework.domain.entity.attribute.Column;
 import is.codion.framework.domain.entity.attribute.ForeignKey;
-import is.codion.framework.domain.entity.attribute.ForeignKeyDefinition;
-import is.codion.framework.domain.entity.attribute.ValueAttributeDefinition;
 import is.codion.framework.model.AbstractEntityEditor;
 import is.codion.framework.model.EntityEditor.EditorTask.Result;
 import is.codion.swing.common.model.component.combobox.SwingComboBoxModel;
@@ -38,11 +33,8 @@ import is.codion.swing.framework.model.component.SwingEntityComboBoxModel;
 
 import org.jspecify.annotations.Nullable;
 
-import java.util.Collection;
-import java.util.HashMap;
 import java.util.Map;
 
-import static java.util.Collections.unmodifiableMap;
 import static java.util.Objects.requireNonNull;
 import static javax.swing.SwingUtilities.isEventDispatchThread;
 
@@ -50,11 +42,6 @@ import static javax.swing.SwingUtilities.isEventDispatchThread;
  * A Swing {@link AbstractEntityEditor} implementation.
  */
 public final class SwingEntityEditor extends AbstractEntityEditor<SwingEntityEditor> {
-
-	private static final String NULL_ITEM_CAPTION = FilterComboBoxModel.NULL_CAPTION.getOrThrow();
-	private static final ProxyBuilder.ProxyMethod<Object> NULL_ITEM_TO_STRING = parameters -> NULL_ITEM_CAPTION;
-
-	private final ComboBoxModels comboBoxModels = new ComboBoxModels();
 
 	private @Nullable ProgressWorker<Result<Entity>, ?> worker;
 	private @Nullable EditorTask<?> currentTask;
@@ -76,14 +63,12 @@ public final class SwingEntityEditor extends AbstractEntityEditor<SwingEntityEdi
 	 */
 	public SwingEntityEditor(EntityType entityType, EntityConnectionProvider connectionProvider,
 													 SwingComponentModels componentModels) {
-		super(entityType, connectionProvider, componentModels);
+		super(entityType, connectionProvider, componentModels, DefaultSwingComboBoxModels::new);
 	}
 
-	/**
-	 * @return the {@link ComboBoxModels} instance
-	 */
-	public ComboBoxModels comboBoxModels() {
-		return comboBoxModels;
+	@Override
+	public SwingComboBoxModels comboBoxModels() {
+		return (SwingComboBoxModels) super.comboBoxModels();
 	}
 
 	@Override
@@ -154,125 +139,42 @@ public final class SwingEntityEditor extends AbstractEntityEditor<SwingEntityEdi
 	/**
 	 * Manages the combo box models used by a {@link SwingEntityEditor}.
 	 */
-	public final class ComboBoxModels {
+	public interface SwingComboBoxModels extends ComboBoxModels {
 
-		private final Map<ForeignKey, SwingEntityComboBoxModel> foreignKeyComboBoxModels = new HashMap<>();
-		private final Map<Column<?>, SwingComboBoxModel<?>> columnComboBoxModels = new HashMap<>();
+		@Override
+		Map<ForeignKey, SwingEntityComboBoxModel> foreignKey();
 
-		private ComboBoxModels() {}
+		@Override
+		Map<Column<?>, SwingComboBoxModel<?>> column();
 
-		/**
-		 * Returns the foreign key based combo box models initialized via {@link #get(ForeignKey)}.
-		 * @return an unmodifiable view of the foreign key based combo box models
-		 */
-		public Map<ForeignKey, SwingEntityComboBoxModel> foreignKey() {
-			synchronized (foreignKeyComboBoxModels) {
-				return unmodifiableMap(foreignKeyComboBoxModels);
-			}
+		@Override
+		SwingEntityComboBoxModel get(ForeignKey foreignKey);
+
+		@Override
+		<T> SwingComboBoxModel<T> get(Column<T> column);
+
+		@Override
+		SwingEntityComboBoxModel create(ForeignKey foreignKey);
+
+		@Override
+		<T> SwingComboBoxModel<T> create(Column<T> column);
+	}
+
+	private static final class DefaultSwingComboBoxModels
+					extends DefaultComboBoxModels<SwingEntityComboBoxModel, SwingComboBoxModel<?>> implements SwingComboBoxModels {
+
+		private DefaultSwingComboBoxModels(AbstractEntityEditor<?> editor) {
+			super(editor);
 		}
 
-		/**
-		 * Returns the column based combo box models initialized via {@link #get(Column)}.
-		 * @return an unmodifiable view of the column based combo box models
-		 */
-		public Map<Column<?>, SwingComboBoxModel<?>> column() {
-			synchronized (columnComboBoxModels) {
-				return unmodifiableMap(columnComboBoxModels);
-			}
-		}
-
-		/**
-		 * Creates and refreshes combo box models for the given attributes. Doing this in a edit model
-		 * constructor avoids the models being refreshed when the combo boxes using them are initialized
-		 * @param attributes the attributes for which to initialize combo box models
-		 * @see #create(Column)
-		 * @see #create(ForeignKey)
-		 */
-		public void initialize(Attribute<?>... attributes) {
-			for (Attribute<?> attribute : requireNonNull(attributes)) {
-				if (attribute instanceof ForeignKey) {
-					get((ForeignKey) attribute).items().refresh();
-				}
-				else if (attribute instanceof Column<?>) {
-					get((Column<?>) attribute).items().refresh();
-				}
-			}
-		}
-
-		/**
-		 * <p>Returns the {@link SwingEntityComboBoxModel} associated with the given foreign key.
-		 * If no such combo box model exists, one is created by calling {@link #create(ForeignKey)}.
-		 * <p>This method always returns the same {@link SwingEntityComboBoxModel} instance, once one has been created.
-		 * @param foreignKey the foreign key
-		 * @return the {@link SwingEntityComboBoxModel} associated with the given foreign key
-		 * @see ComboBoxModels#create(ForeignKey)
-		 */
-		public SwingEntityComboBoxModel get(ForeignKey foreignKey) {
-			entityDefinition().foreignKeys().definition(foreignKey);
-			synchronized (foreignKeyComboBoxModels) {
-				// can't use computeIfAbsent() here, since that prevents recursive initialization of interdepending combo
-				// box models, create() may for example call this function
-				// see javadoc: must not attempt to update any other mappings of this map
-				SwingEntityComboBoxModel comboBoxModel = foreignKeyComboBoxModels.get(foreignKey);
-				if (comboBoxModel == null) {
-					comboBoxModel = create(requireNonNull(foreignKey));
-					foreignKeyComboBoxModels.put(foreignKey, comboBoxModel);
-				}
-
-				return comboBoxModel;
-			}
-		}
-
-		/**
-		 * <p>Returns the {@link SwingComboBoxModel} associated with the given column.
-		 * If no such combo box model exists, one is created by calling {@link #create(Column)}.
-		 * <p>This method always returns the same {@link SwingComboBoxModel} instance, once one has been created.
-		 * @param column the column
-		 * @param <T> the value type
-		 * @return the {@link SwingComboBoxModel} associated with the given column
-		 * @see #create(Column)
-		 */
+		@Override
 		public <T> SwingComboBoxModel<T> get(Column<T> column) {
-			entityDefinition().columns().definition(column);
-			synchronized (columnComboBoxModels) {
-				// can't use computeIfAbsent() here, since that prevents recursive initialization of interdepending combo
-				// box models, create() may for example call this function
-				// see javadoc: must not attempt to update any other mappings of this map
-				SwingComboBoxModel<T> comboBoxModel = (SwingComboBoxModel<T>) columnComboBoxModels.get(column);
-				if (comboBoxModel == null) {
-					comboBoxModel = create(requireNonNull(column));
-					columnComboBoxModels.put(column, comboBoxModel);
-				}
-
-				return comboBoxModel;
-			}
+			return (SwingComboBoxModel<T>) super.get(column);
 		}
 
-		/**
-		 * <p>Creates a new {@link SwingEntityComboBoxModel} for the given foreign key.
-		 * @param foreignKey the foreign key for which to create a {@link SwingEntityComboBoxModel}
-		 * @return a {@link SwingEntityComboBoxModel} for the given foreign key
-		 * @see SwingComponentModels#comboBoxModel(ForeignKey, EntityConnectionProvider)
-		 * @see SwingComboBoxModel#NULL_CAPTION
-		 * @see SwingEntityComboBoxModel.Builder#nullCaption(String)
-		 * @see SwingEntityComboBoxModel.Builder#includeNull(boolean)
-		 * @see ValueAttributeDefinition#nullable()
-		 * @see SwingEntityComboBoxModel.Builder#attributes(Collection)
-		 * @see ForeignKeyDefinition#attributes()
-		 */
-		public SwingEntityComboBoxModel create(ForeignKey foreignKey) {
-			return componentModels().comboBoxModel(requireNonNull(foreignKey), connectionProvider());
-		}
-
-		/**
-		 * Creates a combo box model containing the current values of the given column.
-		 * @param column the column
-		 * @param <T> the value type
-		 * @return a combo box model based on the given column
-		 * @see SwingComboBoxModel#NULL_CAPTION
-		 */
+		@Override
 		public <T> SwingComboBoxModel<T> create(Column<T> column) {
-			return componentModels().comboBoxModel(requireNonNull(column), connectionProvider());
+			return (SwingComboBoxModel<T>) super.create(column);
 		}
 	}
 
@@ -283,39 +185,15 @@ public final class SwingEntityEditor extends AbstractEntityEditor<SwingEntityEdi
 	 */
 	public interface SwingComponentModels extends ComponentModels {
 
-		/**
-		 * <p>Creates a new {@link SwingEntityComboBoxModel} for the given foreign key, override to
-		 * provide a custom {@link SwingEntityComboBoxModel} implementation.
-		 * <p>This default implementation returns a sorted {@link SwingEntityComboBoxModel} using the default
-		 * null item caption if the underlying attribute is nullable.
-		 * <p>If the foreign key has select attributes defined, those are set in the combo box model.
-		 * @param foreignKey the foreign key for which to create a {@link SwingEntityComboBoxModel}
-		 * @param connectionProvider the connection provider
-		 * @return a {@link SwingEntityComboBoxModel} for the given foreign key
-		 * @see SwingComboBoxModel#NULL_CAPTION
-		 * @see SwingEntityComboBoxModel.Builder#nullCaption(String)
-		 * @see SwingEntityComboBoxModel.Builder#includeNull(boolean)
-		 * @see ValueAttributeDefinition#nullable()
-		 * @see SwingEntityComboBoxModel.Builder#attributes(Collection)
-		 * @see ForeignKeyDefinition#attributes()
-		 */
+		@Override
 		default SwingEntityComboBoxModel comboBoxModel(ForeignKey foreignKey, EntityConnectionProvider connectionProvider) {
 			return SwingEntityComboBoxModel.builder()
 							.foreignKey(foreignKey)
-							.connectionProvider(requireNonNull(requireNonNull(connectionProvider)))
+							.connectionProvider(requireNonNull(connectionProvider))
 							.build();
 		}
 
-		/**
-		 * Creates a combo box model containing the current values of the given column.
-		 * This default implementation returns a sorted {@link SwingComboBoxModel} using the default
-		 * null item caption if the underlying column is nullable
-		 * @param column the column
-		 * @param connectionProvider the connection provider
-		 * @param <T> the value type
-		 * @return a combo box model based on the given column
-		 * @see SwingComboBoxModel#NULL_CAPTION
-		 */
+		@Override
 		default <T> SwingComboBoxModel<T> comboBoxModel(Column<T> column, EntityConnectionProvider connectionProvider) {
 			EntityDefinition entityDefinition = requireNonNull(connectionProvider).entities()
 							.definition(requireNonNull(column).entityType());
@@ -323,15 +201,9 @@ public final class SwingEntityEditor extends AbstractEntityEditor<SwingEntityEdi
 
 			return SwingComboBoxModel.builder()
 							.items(() -> connectionProvider.connection().select(column))
-							.nullItem(nullable ? createNullItem(column) : null)
+							.nullItem(nullable ? ComponentModels.createNullItem(column) : null)
 							.includeNull(nullable)
 							.build();
 		}
-	}
-
-	static <T> @Nullable T createNullItem(Column<T> column) {
-		return column.type().valueClass().isInterface() ? ProxyBuilder.of(column.type().valueClass())
-						.method("toString", (ProxyBuilder.ProxyMethod<T>) NULL_ITEM_TO_STRING)
-						.build() : null;
 	}
 }